Bug Fix: Wrong Data Type for Metadata Key Field

+++ src/main/java/rest/koios/client/backend/api/transactions/model/Metadata.java
@@ -21,7 +21,7 @@ public class Metadata {
     /**
      * Metadata key (index)
      */
-    private Integer key;
+    private Long key;
 
     /**
      * A JSON containing details about metadata within a transaction
